move files back to original package
