+++ x-pack/plugin/security/src/main/java/org/elasticsearch/xpack/core/security/transport/SecurityTransportExceptionHandler.java
@@ -4,12 +4,13 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-package org.elasticsearch.xpack.core.security.transport.netty4;
+package org.elasticsearch.xpack.core.security.transport;
 
 import org.apache.logging.log4j.Logger;
 import org.elasticsearch.common.component.Lifecycle;
 import org.elasticsearch.common.network.CloseableChannel;
 import org.elasticsearch.transport.TcpChannel;
+import org.elasticsearch.xpack.core.security.transport.SSLExceptionHelper;
 
 import java.util.function.BiConsumer;
